feat: Add multiple invalidations based on stages
Before this commit, `cloudfrontInvalidation` was a single object under the
`custom` section of serverless.yml file. There was no possible way to create
invalidations for multiple CloudFront distributions. Therefore, the
`cloudfrontInvalidation` object is turned into a list that can contain
multiple distributions. index.js file is modified to realize this feature.
The function body in `invalidate()` is moved into
`...cloudfrontInvalidation.forEach()`

Besides, an option of create invalidations based on deployment stages
was also not available. Hence, an optional `stage` property is added for
each distribution element. An if statement that checks for `stage` property
declaration and its matching with the provider's stage is added into
`invalidate()` in index.js file.

The Setup example of `cloudfrontInvalidation` as a list and the Setup
example of using the `stage` property of the elements of
`cloudfrontInvalidation` is added to README.md file.

Modified files:
    - index.js
    - README.md